+++ src/pyfhd/pyfhd_tools/pyfhd_setup.py
@@ -38,8 +38,19 @@ class OrderedBooleanOptionalAction(argparse.BooleanOptionalAction):
         self.option_strings = longs + shorts
 
 
-def git_info():
-    version_str = version("pyfhd")
+def git_info(version_str: str | None = None):
+    """
+    Parse the version info.
+
+    Parameters
+    ----------
+    version_str : str, optional
+        Version string to parse. If None, get the pyfhd version. Should only be
+        set to a string for testing purposes.
+
+    """
+    if version_str is None:
+        version_str = version("pyfhd")
     parts = version_str.split(".")
 
     if ".dev" not in version_str:
@@ -1002,6 +1013,14 @@ def pyfhd_parser():
     #     action=OrderedBooleanOptionalAction,
     #     help="Run Fast Holographic Deconvolution",
     # )
+    gridding.add_argument(
+        "--recalculate-mapfn",
+        default=False,
+        action=OrderedBooleanOptionalAction,
+        help="Forces pyfhd to recalculate the holographic mapping function. "
+        "Replaces mapfn_recalculate from FHD",
+    )
+
     # deconv.add_argument(
     #     "--max-deconvolution-components",
     #     type=int,
@@ -1736,13 +1755,15 @@ def pyfhd_setup(options: argparse.Namespace) -> Tuple[dict, logging.Logger]:
     if pyfhd_config["recalculate_all"]:
         pyfhd_config["recalculate_beam"] = True
         pyfhd_config["recalculate_grid"] = True
-        pyfhd_config["recalculate_mapfn"] = True
+        # if pyfhd_config["deconvolution"]:
+        #     pyfhd_config["recalculate_mapfn"] = True
         logger.info(
             "Recalculate All option has been enabled, the beam, gridding and map "
             "function will be recalculated"
         )
 
     # If both mapping function and healpix export are on save the visibilities (Warning)
+    # TODO: figure out if this should change now that we can build the mapping function
     if (
         pyfhd_config["snapshot_healpix_export"]
         and not pyfhd_config["save_visibilities"]
@@ -1751,6 +1772,16 @@ def pyfhd_setup(options: argparse.Namespace) -> Tuple[dict, logging.Logger]:
         logger.warning(
             "If we're exporting healpix we should also save the visibilities "
             "that created them. Setting save_visibilities to True"
+        )
+        warnings += 1
+
+    # Turn off calculating the mapping function if grid_uniform is on (warning)
+    # TODO: add deconvolution here too later
+    if pyfhd_config["grid_uniform"] and pyfhd_config["recalculate_mapfn"]:
+        pyfhd_config["recalculate_mapfn"] = False
+        logger.warning(
+            "The grid_uniform and recalculate_mapfn options are incompatible. "
+            "Setting recalculate_mapfn to False."
         )
         warnings += 1